ProductManager 분리
